# Initial, cleaned-up cut of performance tuning patches.