--- a/src/Engines/Nunjucks.js
+++ b/src/Engines/Nunjucks.js
@@ -21,7 +21,7 @@ class Nunjucks extends TemplateEngine {
       new NunjucksLib.Environment(
         new NunjucksLib.FileSystemLoader([
           super.getIncludesDir(),
-          TemplatePath.getWorkingDir()
+          TemplatePath.getWorkingDir(),
         ])
       );
     this.setEngineLib(this.njkEnv);
@@ -89,7 +89,7 @@ class Nunjucks extends TemplateEngine {
     function ShortcodeFunction() {
       this.tags = [shortcodeName];
 
-      this.parse = function(parser, nodes, lexer) {
+      this.parse = function (parser, nodes, lexer) {
         let args;
         let tok = parser.nextToken();
 
@@ -109,7 +109,7 @@ class Nunjucks extends TemplateEngine {
         return new nodes.CallExtension(this, "run", args);
       };
 
-      this.run = function(...args) {
+      this.run = function (...args) {
         let resolve;
         if (isAsync) {
           resolve = args.pop();
@@ -120,10 +120,10 @@ class Nunjucks extends TemplateEngine {
         if (isAsync) {
           shortcodeFn
             .call(Nunjucks._normalizeShortcodeContext(context), ...argArray)
-            .then(function(returnValue) {
+            .then(function (returnValue) {
               resolve(null, new NunjucksLib.runtime.SafeString(returnValue));
             })
-            .catch(function(e) {
+            .catch(function (e) {
               resolve(
                 new EleventyShortcodeError(
                   `Error with Nunjucks shortcode \`${shortcodeName}\`${EleventyErrorUtil.convertErrorToString(
@@ -160,7 +160,7 @@ class Nunjucks extends TemplateEngine {
     function PairedShortcodeFunction() {
       this.tags = [shortcodeName];
 
-      this.parse = function(parser, nodes, lexer) {
+      this.parse = function (parser, nodes, lexer) {
         var tok = parser.nextToken();
 
         var args = parser.parseSignature(true, true);
@@ -175,7 +175,7 @@ class Nunjucks extends TemplateEngine {
         return new nodes.CallExtension(this, "run", args, [body]);
       };
 
-      this.run = function(...args) {
+      this.run = function (...args) {
         let resolve;
         if (isAsync) {
           resolve = args.pop();
@@ -190,10 +190,10 @@ class Nunjucks extends TemplateEngine {
               body(),
               ...argArray
             )
-            .then(function(returnValue) {
+            .then(function (returnValue) {
               resolve(null, new NunjucksLib.runtime.SafeString(returnValue));
             })
-            .catch(function(e) {
+            .catch(function (e) {
               resolve(
                 new EleventyShortcodeError(
                   `Error with Nunjucks paired shortcode \`${shortcodeName}\`${EleventyErrorUtil.convertErrorToString(
@@ -227,9 +227,20 @@ class Nunjucks extends TemplateEngine {
   }
 
   async compile(str, inputPath) {
-    let needsCompile = (str.indexOf("{{") != -1) || (str.indexOf("{%") != -1);
+    // Defend against syntax customisations:
+    //    https://mozilla.github.io/nunjucks/api.html#customizing-syntax
+    let optsTags = this.njkEnv.opts.tags || {};
+    let blockStart = optsTags.blockStart || "{%";
+    let variableStart = optsTags.variableStart || "{{";
+    let commentStart = optsTags.variableStart || "{#";
+    let needsCompile =
+      str.indexOf(blockStart) != -1 ||
+      str.indexOf(variableStart) != -1 ||
+      str.indexOf(commentStart) != -1;
     if (!needsCompile) {
-      return async function() { return str; };
+      return async function () {
+        return str;
+      };
     }
 
     let tmpl;
@@ -238,10 +249,9 @@ class Nunjucks extends TemplateEngine {
     } else {
       tmpl = NunjucksLib.compile(str, this.njkEnv, inputPath);
     }
-    return async function(data) {
-
-      return new Promise(function(resolve, reject) {
-        tmpl.render(data, function(err, res) {
+    return async function (data) {
+      return new Promise(function (resolve, reject) {
+        tmpl.render(data, function (err, res) {
           if (err) {
             reject(err);
           } else {