--- a/src/Engines/TemplateEngine.js
+++ b/src/Engines/TemplateEngine.js
@@ -12,6 +12,7 @@ class TemplateEngine {
     this.partialsHaveBeenCached = false;
     this.partials = [];
     this.engineLib = null;
+    this.cacheable = false;
   }
 
   get config() {