--- a/src/Plugins/Pagination.js
+++ b/src/Plugins/Pagination.js
@@ -113,7 +113,7 @@ class Pagination {
       keys = Object.keys(fullDataSet);
     }
 
-    let result = keys.filter(() => true);
+    let result = keys.slice();
 
     if (
       this.data.pagination.before &&
@@ -128,7 +128,7 @@ class Pagination {
     }
 
     if (this.data.pagination.filter) {
-      result = result.filter(value => !this.isFiltered(value));
+      result = result.filter((value) => !this.isFiltered(value));
     }
 
     return result;
@@ -194,7 +194,7 @@ class Pagination {
           size: this.data.pagination.size,
           alias: this.alias,
 
-          pages: this.size === 1 ? items.map(entry => entry[0]) : items,
+          pages: this.size === 1 ? items.map((entry) => entry[0]) : items,
 
           // See Issue #345 for more examples
           page: {
@@ -219,12 +219,12 @@ class Pagination {
               ? this.size === 1
                 ? items[items.length - 1][0]
                 : items[items.length - 1]
-              : null
+              : null,
           },
 
           items: items[pageNumber],
-          pageNumber: pageNumber
-        }
+          pageNumber: pageNumber,
+        },
       };
 
       if (this.alias) {
@@ -245,7 +245,7 @@ class Pagination {
 
     // we loop twice to pass in the appropriate prev/next links (already full generated now)
     templates.forEach(
-      function(cloned, pageNumber) {
+      function (cloned, pageNumber) {
         let pageObj = {};
 
         // links are okay but hrefs are better
@@ -282,7 +282,7 @@ class Pagination {
           previous: pageObj.previousPageHref,
           next: pageObj.nextPageHref,
           first: pageObj.firstPageHref,
-          last: pageObj.lastPageHref
+          last: pageObj.lastPageHref,
         };
 
         Object.assign(overrides[pageNumber].pagination, pageObj);