--- a/src/TemplateCollection.js
+++ b/src/TemplateCollection.js
@@ -12,7 +12,7 @@ class TemplateCollection extends Sortable {
   }
 
   getAll() {
-    return this.items.filter(() => true);
+    return this.items.slice();
   }
 
   getAllSorted() {
@@ -28,7 +28,7 @@ class TemplateCollection extends Sortable {
       globs = [globs];
     }
 
-    globs = globs.map(glob => TemplatePath.addLeadingDotSlash(glob));
+    globs = globs.map((glob) => TemplatePath.addLeadingDotSlash(glob));
 
     return globs;
   }
@@ -42,14 +42,14 @@ class TemplateCollection extends Sortable {
     if (!this._dirty) {
       // Try to find a pre-sorted list and clone it.
       if (this._filteredByGlobsCache.has(key)) {
-        return [...(this._filteredByGlobsCache.get(key))];
+        return [...this._filteredByGlobsCache.get(key)];
       }
     } else if (this._filteredByGlobsCache.size) {
       // Blow away cache
       this._filteredByGlobsCache = new Map();
     }
 
-    let filtered = this.getAllSorted().filter(item => {
+    let filtered = this.getAllSorted().filter((item) => {
       if (multimatch([item.inputPath], globs).length) {
         return true;
       }
@@ -62,19 +62,19 @@ class TemplateCollection extends Sortable {
   }
 
   getFilteredByTag(tagName) {
-    return this.getAllSorted().filter(item => {
+    return this.getAllSorted().filter((item) => {
       if (!tagName) {
         return true;
       } else if (Array.isArray(item.data.tags)) {
-        return item.data.tags.some(tag => tag === tagName);
+        return item.data.tags.some((tag) => tag === tagName);
       }
       return false;
     });
   }
 
   getFilteredByTags(...tags) {
-    return this.getAllSorted().filter(item =>
-      tags.every(requiredTag => {
+    return this.getAllSorted().filter((item) =>
+      tags.every((requiredTag) => {
         const itemTags = item.data.tags;
         if (Array.isArray(itemTags)) {
           return itemTags.includes(requiredTag);