--- a/src/TemplateData.js
+++ b/src/TemplateData.js
@@ -20,6 +20,22 @@ const deleteRequireCache = require("./Util/DeleteRequireCache");
 const bench = require("./BenchmarkManager").get("Data");
 const aggregateBench = require("./BenchmarkManager").get("Aggregate");
 
+class FSExistsCache {
+  _cache = new Map();
+  has(path) { return this._cache.has(path); }
+  exists(path) {
+    let exists = this._cache.get(path);
+    if (!this.has(path)) {
+      exists = fs.pathExistsSync(path);
+      this._cache.set(path, exists);
+    }
+    return exists;
+  }
+  markExists(path, value=true) {
+    this._cache.set(path, !!value);
+  }
+}
+
 class TemplateDataParseError extends EleventyBaseError {}
 
 class TemplateData {
@@ -104,7 +120,7 @@ class TemplateData {
 
   async _checkInputDir() {
     if (this.inputDirNeedsCheck) {
-      let globalPathStat = await fs.stat(this.inputDir);
+      let globalPathStat = fs.statSync(this.inputDir);
 
       if (!globalPathStat.isDirectory()) {
         throw new Error("Could not find data path directory: " + this.inputDir);
@@ -180,7 +196,7 @@ class TemplateData {
 
     let fsBench = aggregateBench.get("Searching the file system");
     fsBench.before();
-    let paths = await fastglob(await this.getGlobalDataGlob(), {
+    let paths = fastglob.sync(await this.getGlobalDataGlob(), {
       caseSensitiveMatch: false,
       dot: true
     });
@@ -261,12 +277,21 @@ class TemplateData {
     return this.globalData;
   }
 
+
   /* Template and Directory data files */
   async combineLocalData(localDataPaths) {
     let localData = {};
     if (!Array.isArray(localDataPaths)) {
       localDataPaths = [localDataPaths];
     }
+
+    // Filter out files we know don't exist to avoid overhead for checking
+    localDataPaths = localDataPaths.filter((path) => {
+      return this._fsExistsCache.exists(path);
+    });
+
+    if (!localDataPaths.length) { return localData; }
+
     for (let path of localDataPaths) {
       // clean up data for template/directory data files only.
       let dataForPath = await this.getDataValue(path, null, true);
@@ -358,6 +383,10 @@ class TemplateData {
     }
   }
 
+  // It's common for data files not to exist, so we avoid going to the FS to
+  // re-check if they do via a quick-and-dirty cache.
+  _fsExistsCache = new FSExistsCache();
+
   async getDataValue(path, rawImports, ignoreProcessing) {
     let extension = TemplatePath.getExtension(path);
 
@@ -370,7 +399,11 @@ class TemplateData {
     ) {
       // JS data file or require’d JSON (no preprocessing needed)
       let localPath = TemplatePath.absolutePath(path);
-      if (!(await fs.pathExists(localPath))) {
+      let exists = this._fsExistsCache.exists(localPath);
+      // Make sure that relative lookups benefit from cache
+      this._fsExistsCache.markExists(path, exists);
+
+      if (!exists) {
         return {};
       }