Benchmark results:

  https://gist.github.com/slightlyoff/4a2c2709c8a4ef80ef074010b0be6a65

API compatible; all tests pass.