@nhoizey 2020-06-20
On my own site, I get a **9 % build time improvement!** 👍

10 builds with 0.11.0:

- 36.68 s
- 32.82 s
- 33.05 s
- 33.77 s
- 33.26 s
- 33.46 s
- 33.16 s
- 32.66 s
- 32.95 s
- 32.99 s

→ mean 33.48 s

10 builds with this PR :

- 30.45 s
- 29.37 s
- 34.73 s
- 32.52 s
- 28.37 s
- 28.56 s
- 29.61 s
- 29.33 s
- 30.67 s
- 30.89 s

→ mean 30.45 s

(and no issue with the result of the build… 😅)