@mathiasbynens 2020-06-22
Just tested this on [the v8.dev repository](https://github.com/v8/v8.dev). Although this is a relatively small site, I saw an overall speedup of 4–5%:

```sh
$ (for i in {1..10}; do npm run eleventy; done) | grep Copied
Copied 496 files / Wrote 249 files in 1.73 seconds (6.9ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.56 seconds (6.3ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.59 seconds (6.4ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.56 seconds (6.3ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.57 seconds (6.3ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.56 seconds (6.3ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.52 seconds (6.1ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.57 seconds (6.3ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.65 seconds (6.6ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.57 seconds (6.3ms each, v0.11.0)

$ npm install slightlyoff/eleventy#faster-builds

$ (for i in {1..10}; do npm run eleventy; done) | grep Copied
Copied 496 files / Wrote 249 files in 1.49 seconds (6.0ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.55 seconds (6.2ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.50 seconds (6.0ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.46 seconds (5.9ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.43 seconds (5.7ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.59 seconds (6.4ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.48 seconds (5.9ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.59 seconds (6.4ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.53 seconds (6.1ms each, v0.11.0)
Copied 496 files / Wrote 249 files in 1.53 seconds (6.1ms each, v0.11.0)
```