@mathiasbynens 2020-06-22
Testing on [the web.dev repository](https://github.com/GoogleChrome/web.dev), after patching its `package.json` to remove [the `--quiet` flag](https://github.com/GoogleChrome/web.dev/blob/ae8d31a3c7c4a98937d0750da37a6f5f8d6ef125/package.json#L17):

```
$ (for i in {1..10}; do npm run eleventy; done) | grep Wrote
Wrote 759 files in 15.32 seconds (20.2ms each, v0.11.0)
Wrote 759 files in 16.90 seconds (22.3ms each, v0.11.0)
Wrote 759 files in 15.28 seconds (20.1ms each, v0.11.0)
Wrote 759 files in 21.60 seconds (28.5ms each, v0.11.0)
Wrote 759 files in 18.23 seconds (24.0ms each, v0.11.0)
Wrote 759 files in 19.13 seconds (25.2ms each, v0.11.0)
Wrote 759 files in 15.44 seconds (20.3ms each, v0.11.0)
Wrote 759 files in 14.24 seconds (18.8ms each, v0.11.0)
Wrote 759 files in 13.42 seconds (17.7ms each, v0.11.0)
Wrote 759 files in 11.31 seconds (14.9ms each, v0.11.0)

$ npm install slightlyoff/eleventy#faster-builds

$ (for i in {1..10}; do npm run eleventy; done) | grep Wrote
Wrote 759 files in 10.33 seconds (13.6ms each, v0.11.0)
Wrote 759 files in 10.07 seconds (13.3ms each, v0.11.0)
Wrote 759 files in 10.26 seconds (13.5ms each, v0.11.0)
Wrote 759 files in 10.16 seconds (13.4ms each, v0.11.0)
Wrote 759 files in 9.86 seconds (13.0ms each, v0.11.0)
Wrote 759 files in 9.93 seconds (13.1ms each, v0.11.0)
Wrote 759 files in 10.80 seconds (14.2ms each, v0.11.0)
Wrote 759 files in 10.26 seconds (13.5ms each, v0.11.0)
Wrote 759 files in 9.98 seconds (13.1ms each, v0.11.0)
Wrote 759 files in 10.24 seconds (13.5ms each, v0.11.0)
```

That's a 34–37% speedup!

- Without patch: median=15.38, average=16.09
- With patch: median=10.20 (+34%), average=10.19 (+37%)