@dgrammatiko 2020-06-22
I have tested this PR with results:
```
Before:
Copied 1520 files / Wrote 772 files in 2.38 seconds (3.1ms each, v0.11.0)
after
Copied 1520 files / Wrote 772 files in 2.21 seconds (2.9ms each, v0.11.0)

```

7.15% less time...

Test site an alteration of https://joomla-needed-fixes.netlify.app/data/ (repo: https://github.com/dgrammatiko/showcase-audit/tree/perf )