@edwardhorsford 2020-08-01
I just tried this on my build.

Build times went from **~3.7s** to **~2.8s** 🎉 

Note, swapping to it *did* crash my build. Doing some digging I _think_ the issue is that at some point I was doing a read of a json file _whilst_ it was still being written by another bit of code. Swapping the write to sync fixed it. I'd guess that previously eleventy was _slow_ enough that the write always completed in time - whereas now it's just too fast.