@zachleat 2020-08-06
@edwardhorsford just to check in—this was in your app code right? You didn’t tweak Eleventy internals?