@edwardhorsford 2020-08-06
@zachleat yes, correct. I *think* the optimisations in this uncovered a bug in my existing conde.